+++ src/devices/ikea.ts
@@ -76,25 +76,25 @@ const configureRemote: Configure = async (device, coordinatorEndpoint, logger) =
 };
 
 const tradfriExtend = {
-    light_onoff_brightness: (options: Extend.options_light_onoff_brightness = {}) => ({
+    light_onoff_brightness: (options: Extend.options_light_onoff_brightness = {}): Extend => ({
         ...extend.light_onoff_brightness(options),
         ota: ota.tradfri,
         onEvent: bulbOnEvent,
     }),
-    light_onoff_brightness_colortemp: (options: Extend.options_light_onoff_brightness_colortemp = {colorTempRange: [250, 454]}) => ({
+    light_onoff_brightness_colortemp: (options: Extend.options_light_onoff_brightness_colortemp = {colorTempRange: [250, 454]}): Extend => ({
         ...extend.light_onoff_brightness_colortemp(options),
         exposes: [...extend.light_onoff_brightness_colortemp(options).exposes, e.light_color_options()],
         ota: ota.tradfri,
         onEvent: bulbOnEvent,
     }),
     light_onoff_brightness_colortemp_color: (
-        options: Extend.options_light_onoff_brightness_colortemp_color = {disableColorTempStartup: true, colorTempRange: [250, 454]}) => ({
+        options: Extend.options_light_onoff_brightness_colortemp_color = {disableColorTempStartup: true, colorTempRange: [250, 454]}): Extend => ({
         ...extend.light_onoff_brightness_colortemp_color(options),
         exposes: [...extend.light_onoff_brightness_colortemp_color(options).exposes, e.light_color_options()],
         ota: ota.tradfri,
         onEvent: bulbOnEvent,
     }),
-    light_onoff_brightness_color: (options: Extend.options_light_onoff_brightness_color = {}) => ({
+    light_onoff_brightness_color: (options: Extend.options_light_onoff_brightness_color = {}): Extend => ({
         ...extend.light_onoff_brightness_color(options),
         exposes: [...extend.light_onoff_brightness_color(options).exposes, e.light_color_options()],
         ota: ota.tradfri,
@@ -1248,6 +1248,26 @@ const definitions: Definition[] = [
         vendor: 'IKEA',
         description: 'ORMANAS LED strip',
         extend: tradfriExtend.light_onoff_brightness_colortemp_color({colorTempRange: [250, 454]}),
+    },
+    {
+        zigbeeModel: ['VALLHORN Wireless Motion Sensor'],
+        model: 'E2134',
+        vendor: 'IKEA',
+        description: 'VALLHORN wireless motion sensor',
+        fromZigbee: [fz.occupancy, fz.battery, fz.illuminance],
+        toZigbee: [],
+        exposes: [e.occupancy(), e.battery(), e.illuminance()],
+        configure: async (device, cordinatorEndpoint, logger) => {
+            const endpoint1 = device.getEndpoint(1);
+            const endpoint2 = device.getEndpoint(2);
+            const endpoint3 = device.getEndpoint(3);
+            await reporting.bind(endpoint1, cordinatorEndpoint, ['genPowerCfg']);
+            await reporting.batteryPercentageRemaining(endpoint1);
+            await reporting.bind(endpoint2, cordinatorEndpoint, ['msOccupancySensing']);
+            await reporting.occupancy(endpoint2);
+            await reporting.bind(endpoint3, cordinatorEndpoint, ['msIlluminanceMeasurement']);
+            await reporting.illuminance(endpoint3);
+        },
     },
 ];